modify HomePage
