+++ src/components/Carousel/Carousel.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import styled from 'styled-components'
 import { Link } from 'react-router-dom'
 import { MEDIA_QUERY_SM, MEDIA_QUERY_MD } from '../../styles/breakpoints'
@@ -6,7 +6,6 @@ import 'slick-carousel/slick/slick.css'
 import 'slick-carousel/slick/slick-theme.css'
 import Slider from 'react-slick'
 import rwdSettings from './rwdSettings'
-import { getAllPosts } from '../../WebAPI'
 
 const Container = styled.div`
   width: 100%;
@@ -94,18 +93,7 @@ const ItemName = styled.p`
   white-space: nowrap;
 `
 
-function Carousel() {
-  const [posts, setPosts] = useState([])
-
-  useEffect(() => {
-    const fetchPosts = async () => {
-      const res = await getAllPosts(12)
-      setPosts(res.data.allPosts)
-    }
-
-    fetchPosts()
-  }, [])
-
+function Carousel({ posts }) {
   return (
     <>
       {posts.length !== 0 && (
